+++ server/startup.js
@@ -2,7 +2,7 @@ import downloader from 'youtube-dl';
 import mpv from 'node-mpv';
 import fs  from 'fs';
 import sanitize from 'sanitize-filename'
-
+import fsw from 'file-size-watcher';
 
 // global variable to access mpv
 mpv_player = null;
@@ -29,6 +29,8 @@ Meteor.startup(function() {
         time_update: 0.5
     });
 
+
+
     // set up the events
     mpv_player.on('started', Meteor.bindEnvironment(mpv_started));
     mpv_player.on('stopped', Meteor.bindEnvironment(mpv_stopped));
@@ -37,10 +39,12 @@ Meteor.startup(function() {
     mpv_player.on('statuschange', Meteor.bindEnvironment(mpv_statuschange));
     mpv_player.on('timeposition', Meteor.bindEnvironment(mpv_timeposition));
 
+
     // create the folder to store the cached songs
     if(!fs.existsSync('songs')){
         fs.mkdirSync('songs');
     }
+
 
     // create the status database object, if not already created
     if(Status.find().count() == 0){
@@ -54,9 +58,11 @@ Meteor.startup(function() {
         Status.update({}, {$set: {currentPosition: 0, playing: false}});
     }
 
+
     // set the volume
     mpv_player.volume(Status.findOne().volume);
     volume_before_mute = Status.findOne().volume;
+
 
     // download every song if cached mode is activated
     if (cached) {
@@ -68,24 +74,58 @@ Meteor.startup(function() {
             // download the song
             var song = downloader(queuedSong.url, args);
 
+            var filename = `${sanitize(queuedSong.title, " ")}.mp3`;
+
             // write it to the HDD
-            song.pipe(fs.createWriteStream(`songs/${sanitize(queuedSong.title, " ")}.mp3`));
+            song.pipe(fs.createWriteStream(`songs/${filename}`));
 
             console.log(`Handling title "${queuedSong.title}"`);
         });
     }
+
 
     // syncs the client time with the real time roughly every 20 seconds
     setInterval(Meteor.bindEnvironment(function(){
         Status.update({}, {$set: {'currentPosition': timeposition}});
     }), 20000);
 
+
     // if there are songs in the queue, play them
     if(Playlist.find().count() > 0){
-        // TODO find something better for the timeout
-        setTimeout(Meteor.bindEnvironment(function(){
-            Meteor.call('play');
-        }), 3000);
+        if(cached){
+
+            var file = Playlist.find({'position': 0}).file;
+
+            // in the rare case, that the server mode was switched to cached, but the remaining
+            // playlist entries are uncached, there won't be any filestring
+            if(file){
+                // check for the filesize
+                if(fs.statSync(file)["size"] > 200000){
+                    Meteor.call('play');
+                }
+                else{
+                    // check for the filesize because when 'play' is called the early, the filesize is so small
+                    // playback stops immediatly
+                    var fileWatcher = fsw.watch(file).on('sizeChange', Meteor.bindEnvironment(function (newSize, oldSize){
+                        // when the file is bigger than 200kB play it
+                        if(newSize > 200000){
+                            Meteor.call('play');
+                            // release the file watcher
+                            fileWatcher.stop();
+                        }
+                    }));   
+                }
+            }
+            // when there is no file just let the 'play' method handle the playback
+            else{
+                Meteor.call('play');
+            }
+                             
+          }
+      // uncached case is just streaming anyway
+      else{
+        Meteor.call('play');
+      }
     }
 
 });